refactor: rest client
