+++ src/rest/v2/oauth2/rest/index.ts
@@ -1,0 +1,27 @@
+export {
+    type PatreonErrorData,
+} from './error'
+
+export {
+    PATREON_RESPONSE_HEADERS,
+    ResponseHeaders,
+    type PatreonHeadersData,
+    type RestHeaders,
+} from './headers'
+
+export {
+    DefaultRestOptions,
+    RequestMethod,
+    type InternalRequestOptions,
+    type RESTOptions,
+    type RequestOptions,
+    type RestEventMap,
+    type RestFetcher,
+    type RestResponse,
+} from './options'
+
+export {
+    type RestRetriesOptions,
+    type RestRetriesBackoffOptions,
+    type RestRetries
+} from './retries'
